Remove big comment if unchanged from filters file and move upgrade functions to new upgrade-steps.js
