+++ migrator/migration-steps.spec.js
@@ -11,17 +11,6 @@ jest.mock('fs-extra', () => {
   }
 })
 
-jest.mock('fs', () => {
-  const readFile = jest.fn().mockResolvedValue(true)
-  const writeFile = jest.fn().mockResolvedValue(true)
-  return {
-    promises: {
-      readFile,
-      writeFile
-    }
-  }
-})
-
 jest.mock('./reporter', () => {
   const mockReporter = jest.fn()
   return {
@@ -54,7 +43,6 @@ jest.mock('./file-helpers', () => {
 })
 
 const fse = require('fs-extra')
-const fsp = require('fs').promises
 const reporter = require('./reporter')
 const fileHelpers = require('./file-helpers')
 const config = require('../lib/config')
@@ -62,7 +50,14 @@ const { projectDir, starterDir, appDir } = require('../lib/utils/paths')
 
 const migrationSteps = require('./migration-steps')
 const { preflightChecks, deleteIfUnchanged, removeOldPatternIncludesFromSassFile } = require('./migration-steps')
-const { migrateConfig, prepareAppRoutes, prepareSass, deleteUnusedFiles, deleteUnusedDirectories, upgradeIfUnchanged, upgradeIfPossible } = migrationSteps
+const {
+  migrateConfig,
+  prepareAppRoutes,
+  prepareSass,
+  deleteUnusedFiles,
+  deleteUnusedDirectories,
+  upgradeIfUnchanged
+} = migrationSteps
 
 describe('migration steps', () => {
   const mockReporter = reporter.mockReporter
@@ -343,196 +338,6 @@ describe('migration steps', () => {
 
     expect(fileHelpers.deleteFile).toHaveBeenCalledTimes(1)
     expect(fileHelpers.deleteFile).toHaveBeenNthCalledWith(1, path.join(projectDir, unbrandedLayout))
-
-    expect(mockReporter).toHaveBeenCalledTimes(1)
-    expect(mockReporter).toHaveBeenCalledWith(true)
-  })
-
-  it('upgrade application file if possible replacing jquery ready', async () => {
-    const applicationJsFile = path.join('app', 'assets', 'javascripts', 'application.js')
-    const matchFound = false
-    const originalFileContents = `/* global $ */
-
-// Warn about using the kit in production
-if (window.console && window.console.info) {
-  window.console.info('GOV.UK Prototype Kit - do not use for production')
-}
-
-$(document).ready(function () {
-  window.GOVUKFrontend.initAll()
-  console.log('Hello')
-})
-`
-    const expectedFileContents = `//
-// For guidance on how to add JavaScript see:
-// https://prototype-kit.service.gov.uk/docs/adding-css-javascript-and-images
-//
-
-window.GOVUKPrototypeKit.documentReady(function () {
-  console.log('Hello')
-})
-`
-
-    fsp.readFile
-      .mockReturnValueOnce(originalFileContents)
-
-    // mock call upgradeIfPossible method (get this working first)
-    const result = await upgradeIfPossible(applicationJsFile, matchFound)
-    expect(result).toBeTruthy()
-    const expectedFileName = path.join(projectDir, applicationJsFile)
-
-    expect(fsp.readFile).toHaveBeenCalledTimes(1)
-    expect(fsp.readFile).toHaveBeenCalledWith(expectedFileName, 'utf8')
-
-    expect(fsp.writeFile).toHaveBeenCalledTimes(1)
-    const [actualFileName, actualFileContent] = fsp.writeFile.mock.calls[0]
-    expect(actualFileName).toEqual(expectedFileName)
-    expect(actualFileContent).toEqual(expectedFileContents)
-
-    expect(reporter.addReporter).toHaveBeenCalledTimes(1)
-    expect(reporter.addReporter).toHaveBeenCalledWith(`Update ${applicationJsFile}`)
-
-    expect(mockReporter).toHaveBeenCalledTimes(1)
-    expect(mockReporter).toHaveBeenCalledWith(true)
-  })
-
-  it('upgrade application file if possible without replacing jquery ready', async () => {
-    const applicationJsFile = path.join('app', 'assets', 'javascripts', 'application.js')
-    const matchFound = false
-    const originalFileContents = `/* global $ */
-
-// Warn about using the kit in production
-if (window.console && window.console.info) {
-  window.console.info('GOV.UK Prototype Kit - do not use for production')
-}
-
-$(document).ready(function () {
-  window.GOVUKFrontend.initAll()
-  $('.my-button').click(() => console.log('clicked'))
-})
-`
-    const expectedFileContents = `/* global $ */
-
-//
-// For guidance on how to add JavaScript see:
-// https://prototype-kit.service.gov.uk/docs/adding-css-javascript-and-images
-//
-
-
-$(document).ready(function () {
-  $('.my-button').click(() => console.log('clicked'))
-})
-`
-
-    fsp.readFile.mockReturnValue(originalFileContents) // For the second call
-
-    // mock call upgradeIfPossible method (get this working first)
-    const result = await upgradeIfPossible(applicationJsFile, matchFound)
-    expect(result).toBeTruthy()
-    const expectedFileName = path.join(projectDir, applicationJsFile)
-
-    expect(fsp.readFile).toHaveBeenCalledTimes(1)
-    expect(fsp.readFile).toHaveBeenCalledWith(expectedFileName, 'utf8')
-
-    expect(fsp.writeFile).toHaveBeenCalledTimes(1)
-    const [actualFileName, actualFileContent] = fsp.writeFile.mock.calls[0]
-    expect(actualFileName).toEqual(expectedFileName)
-    expect(actualFileContent).toEqual(expectedFileContents)
-
-    expect(reporter.addReporter).toHaveBeenCalledTimes(1)
-    expect(reporter.addReporter).toHaveBeenCalledWith(`Update ${applicationJsFile}`)
-
-    expect(mockReporter).toHaveBeenCalledTimes(1)
-    expect(mockReporter).toHaveBeenCalledWith(true)
-  })
-
-  it('upgrade filters file if possible', async () => {
-    const filtersJsFile = path.join('app', 'assets', 'javascripts', 'filters.js')
-    const matchFound = false
-    const originalFileContents = `module.exports = function (env) {
-  var filters = {}
-
-  /* ------------------------------------------------------------------
-    add your methods to the filters obj below this comment block:
-    @example:
-
-    filters.sayHi = function(name) {
-        return 'Hi ' + name + '!'
-    }
-    
-    For more on filters and how to write them see the Nunjucks
-    documentation.
-
-  ------------------------------------------------------------------ */
-
-  function getData(){
-    var dummyData = require('./data/data.js')
-    return dummyData;
-  }
-
-  filters.toID = function (str) {
-    return str.replaceAll(" ", "-").toLowerCase();
-  }
-
-  /* ------------------------------------------------------------------
-    keep the following line to return your filters to the app
-  ------------------------------------------------------------------ */
-  return filters
-}
-`
-    const starterFileContents = `const govukPrototypeKit = require('govuk-prototype-kit')
-const addFilter = govukPrototypeKit.views.addFilter
-`
-    const expectedFileContents = `const govukPrototypeKit = require('govuk-prototype-kit')
-const addFilter = govukPrototypeKit.views.addFilter
-
-  var filters = {}
-
-  /* ------------------------------------------------------------------
-    add your methods to the filters obj below this comment block:
-    @example:
-
-    filters.sayHi = function(name) {
-        return 'Hi ' + name + '!'
-    }
-    
-    For more on filters and how to write them see the Nunjucks
-    documentation.
-
-  ------------------------------------------------------------------ */
-
-  function getData(){
-    var dummyData = require('./data/data.js')
-    return dummyData;
-  }
-
-  filters.toID = function (str) {
-    return str.replaceAll(" ", "-").toLowerCase();
-  }
-
-// Add the filters using the addFilter function
-Object.entries(filters).forEach(([name, fn]) => addFilter(name, fn))
-`
-
-    fsp.readFile
-      .mockReturnValueOnce(originalFileContents) // For the first call
-      .mockReturnValueOnce(starterFileContents) // For the second call
-
-    // mock call upgradeIfPossible method (get this working first)
-    const result = await upgradeIfPossible(filtersJsFile, matchFound)
-    expect(result).toBeTruthy()
-    const expectedFileName = path.join(projectDir, filtersJsFile)
-
-    expect(fsp.readFile).toHaveBeenCalledTimes(2)
-    expect(fsp.readFile).toHaveBeenCalledWith(expectedFileName, 'utf8')
-
-    expect(fsp.writeFile).toHaveBeenCalledTimes(1)
-    const [actualFileName, actualFileContent] = fsp.writeFile.mock.calls[0]
-    expect(actualFileName).toEqual(expectedFileName)
-    expect(actualFileContent).toEqual(expectedFileContents)
-
-    expect(reporter.addReporter).toHaveBeenCalledTimes(1)
-    expect(reporter.addReporter).toHaveBeenCalledWith(`Update ${filtersJsFile}`)
 
     expect(mockReporter).toHaveBeenCalledTimes(1)
     expect(mockReporter).toHaveBeenCalledWith(true)
